DistanceWarning working (with fake gps)

+++ app/src/main/java/com/example/ss18/msp/lmu/msp_projectkickoff_ss188/Utility/Constants.java
@@ -1,5 +1,5 @@
 package com.example.ss18.msp.lmu.msp_projectkickoff_ss188.Utility;
 
 public final class Constants {
-    public static final int MAX_GPS_DISTANCE = 10;
+    public static final float MAX_GPS_DISTANCE = 10.f;
 }
